Organize services/queries directory the same way I've organized the covid queries
