+++ client/src/models/services/queries/index.ts
@@ -1,0 +1,11 @@
+export {
+  promisedServicesForOrg,
+  suspendedServicesForOrg,
+  useServicesForOrg,
+} from "./ServicesForOrg/ServicesForOrg";
+
+export {
+  promisedServicesForProgram,
+  suspendedServicesForProgram,
+  useServicesForProgram,
+} from "./ServicesForProgram/ServicesForProgram";
